refactor: 클래스, 함수, 변수명 변경
