+++ src/main/kotlin/model/result/PlayerProfit.kt
@@ -1,0 +1,3 @@
+package model.result
+
+data class PlayerProfit(val name: String, val profit: Float)
