--- src/main/kotlin/model/result/PlayerResult.kt
@@ -1,3 +1,0 @@
-package model.result
-
-data class PlayerResult(val name: String, val profit: Float)
